Only visit types with parent ATs when inheriting
When inheriting method ATs, classes may need to be processed that are
not found in the source AT set as they inherit their ATs from their
parent.
The previous iteration would accomplish this by simply visiting the
entire source root.

The commit introduces a simplistic heuristic, filtering out any types
from processing that do not have a parent with an AT defined.
While this is not exact, e.g. a class will still be processed if its
parent type has a field changed via AT, the accesstransformer library
used does not expose the types of ATs defined on a class, making this a
best effort approximation.